FEATURE: Updated the code for the getOptimizedPicture function
The function get the an item from a hash provided in path parameters.

If the optimized cell is empty for a given hash, it means that the optimized image isn't generated yet. We generate it and send it to the user as an image. We also update the DynamoDB item.

If the optimized image is already generated, we just send it to the user.

Co-authored : @callistas