FEATURE: Updated the code for the HandleFileAddedInS3 function
We now verify the image type and only accept the ones that are allowed by the sharp module.

We added the optimized and the baseUrlAWS data in the DynamoDB item in order to know if an optimized image was already generated and to keep the url of aws dynamically given when an image is created (depending the bucket name).

Co-authored : @callistas